Created a more efficient Hall class.
A hall is now a pair of halls, a vertical and horizontal hall that
connect to make a path between two rooms. When a hall is created with
the public constructor class, it reads in the starting point from room 1
to the ending point in room 2, as well as if the initial path will be
vertical or horizontal(arbitrary for now). The class then determines a
point that is a straight line away from both the starting point and the
ending point, and creates a horizontal and vertical path connecting both
rooms.