Updated overview at top of RoomGenerator.java for new version
When RoomGenerator gets called, it creates a char[][] house where 'O'
denotes a walkable field and 'X' denotes a wall. 
  
Call RoomGenerator(width,height) to denote the width and height of the
map in tiles. Then, call RoomGenerator.getMap() to return the char[][]
house map when done.

Currently the halls are working better than previous version. Program
isn't liable to crash anymore (as far as I can tell). 20x20 is probably
the smallest map to go with, though.
  
Rooms are not guaranteed to be connecting.
  
Still only makes 5 rooms. Subject to change.