Updates makeHalls method and addHallToMap for use with new Room class
Now uses a boolean isConnected from the Room class to only attach
hallways from rooms to rooms that are already connected, to try and make
every room reachable.

Changes the addHallToMap method to properly build a path from one point
to the next. Prior versions wouldn't build a path from right to left or
south to north. Now properly builds from right to left or left to right,
same with up and down.

Weird bug still happens that I can't manage to figure out - sometimes a
room is still not connected, but most the time a map is generated just
fine.