Updated makeHalls and addHallsToMap to work with updated Rooms class
When build halls, now only paths from the currentRoom selected to a room
already pathable will be created. This makes it so (well, it should)
every room will be connected. Still a bug that occurs maybe 20% of the
time happens where a room isn't pathable, though.

Also addHallsToMap now properly makes a path regardless of if the
destination point is west or north from the original position. Before,
only a path from west to east or north to south would be made. Now we
can build paths from west to east or east to west, same with north and
south.