Creates isPointLegal method to test legal positions
I think this will be very useful for collision testing. Essentially, you
pass the (players position)/tile_size to get the players location
represented on the 2D array. If the players position does not equal 'X',
a wall, then the position is valid. 

There's still a little bit more thinking to go into it but I think this
is the route that I was planning to use in conjunction with the Zombie
AI and I feel this logic could also be useful with the camera/player
collision testing