Makes cylinder radius half the size.
Collision points are adjusted accordingly. This allows you to have a
better idea what the bug currently is. Right now if I had to guess,
somewhere in the creating of the map it misaligns slightly with the 2d
array of the map's tiles. So the location of each tile in the 2d array
is misalligned with the graphical representation of the tile.

Total speculation though. I don't really know.