Changes the generateHitbox method in Hitbox.java to take in a Group
Changed to a group as the Zombie model is in a group and not an xform. 

A group is a more general Xform. Not definitely sure if this change made
any notable differences. I don't think it does, but if anything starts
acting funny then this might be a good starting point to look at.