Zombies now announce to the console when they MAY be in smell range.
Zombies check the euclidean distance every 2 seconds to see if they MAY
be close enough to smell the player. If the euclidean distance is larger
than the smell range then there is no chance that the zombie may smell
the player, so we don't waste time checking deeper. Else, we use the
pathfinder to see if there exists a path within the smell range.