All zombies are now of type OurZombie.
Fix for a strange problem that would lead to errors in map making. The
problem must be somewhere in the abstract class and how the subclasses
interact with it. All zombies will be of the same type of class, but
will have a boolean denoting whether it is a linewalk or randomwalk, and
then behaving as such.